dfa-DS/requirements.txt
joblib
langchain
langchain_google_genai
langchain_core
python-docx
pypdf2